Refactor: auth를 관리하는 Service 클래스로 역할 분리
